Add model configuration with full-stack CRUD and connection testing
- Backend: ModelConfig model, schema, service, and REST API for CRUD, activation, and provider connection testing
- Electron: IPC handlers for PUT and DELETE methods; keystore integration for API key storage
- Frontend: SettingsModel.vue page with config list, create/edit form, key management dialog, and connection testing UI
- Migration: bump db version to 2, run create_all on version match to discover new tables

+++ backend/app/models/audit.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from datetime import datetime
 
-from sqlalchemy import BigInteger, Column, DateTime, Integer, String, Text, func
+from sqlalchemy import Column, DateTime, Integer, String, Text, func
 
 from app.core.database import Base
 
@@ -14,10 +14,10 @@ class AuditLog(Base):
 
     __tablename__ = "audit_logs"
 
-    id = Column(BigInteger, primary_key=True, autoincrement=True, comment="主键 ID")
+    id = Column(Integer, primary_key=True, autoincrement=True, comment="主键 ID")
     action = Column(String(64), nullable=False, index=True, comment="操作类型，如 'memory.delete', 'model.config.save'")
     target_type = Column(String(64), nullable=True, index=True, comment="操作对象类型，如 'memory', 'session'")
-    target_id = Column(BigInteger, nullable=True, comment="操作对象 ID")
+    target_id = Column(Integer, nullable=True, comment="操作对象 ID")
     summary = Column(String(256), nullable=True, comment="操作摘要（不含敏感正文）")
     detail = Column(Text, nullable=True, comment="操作补充信息（JSON 字符串，不含敏感正文）")
     result = Column(Integer, nullable=False, default=0, comment="操作结果，0=成功，1=失败")
